requirements.txt
torch>=2.2,<2.11
torchvision>=0.17
transformers>=4.40,<5.6
Pillow>=10.0
numpy>=1.26,<3.0
coremltools>=8.0,<10.0
pycocotools>=2.0.7
tqdm>=4.66
matplotlib>=3.8
